ocr_service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pillow==11.1.0
pix2tex==0.1.2
albumentations<2.0
